+++ R/seqArchRplus_generate_reports.R
@@ -1,3 +1,0 @@
-## seqArchRplus generate reports
-##
-